+++ src/System/Symlink.php
@@ -20,53 +20,8 @@ use Composer\Util\Filesystem;
  * @author Antonio J. García Lagar <aj@garcialagar.es>
  */
 class Symlink {
-    public function create(Event $event)
-    {
-        $symlinks = $this->get( $event );
 
-        $fs = new Filesystem();
 
-        foreach ( $event->getComposer()->getRepositoryManager()->getLocalRepository()->getPackages() as $package ) {
-
-            if ( isset( $symlinks[$package->getName()] ) ) {
-
-                $packageDir = $event->getComposer()->getInstallationManager()->getInstallPath( $package );
-
-                $symlinkDefinitions = $symlinks[$package->getName()];
-
-                foreach ( $symlinkDefinitions as $target => $link ) {
-
-                    if ( $fs->isAbsolutePath( $target ) ) {
-
-                        throw new \InvalidArgumentException( "Invalid symlink target path '$target' for package'{$package->getName()}'." . ' It must be relative.' );
-                    }
-
-                    if ( $fs->isAbsolutePath( $link ) ) {
-
-                        throw new \InvalidArgumentException( "Invalid symlink link path '$link' for package'{$package->getName()}'." . ' It must be relative.' );
-                    }
-
-                    $targetPath = $packageDir . DIRECTORY_SEPARATOR . $target;
-                    $linkPath   = getcwd() . DIRECTORY_SEPARATOR . $link;
-
-                    if ( !file_exists( $targetPath ) ) {
-
-                        throw new \RuntimeException( "The target path '$targetPath' for package'{$package->getName()}' does not exist." );
-                    }
-
-                    if ( !file_exists( $linkPath ) ) {
-
-                        $event->getIO()
-                              ->write( sprintf( "  Symlinking <comment>%s</comment> to <comment>%s</comment>", str_replace( getcwd(), '', $targetPath ), str_replace( getcwd(), '', $linkPath ) ) );
-
-                        $fs->ensureDirectoryExists( dirname( $linkPath ) );
-                        $fs->relativeSymlink( $targetPath, $linkPath );
-
-                    }
-                }
-            }
-        }
-    }
 
     protected function get(Event $event)
     {